+++ accounting/invoice.go
@@ -46,7 +46,7 @@ type Invoice struct {
 	CurrencyCode string `json:"CurrencyCode,omitempty" xml:"CurrencyCode,omitempty"`
 
 	// The currency rate for a multicurrency invoice. If no rate is specified, the XE.com day rate is used. (max length = [18].[6])
-	CurrencyRate float32 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
+	CurrencyRate float64 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
 
 	// See Invoice Status Codes
 	Status string `json:"Status,omitempty" xml:"Status,omitempty"`
@@ -61,16 +61,16 @@ type Invoice struct {
 	PlannedPaymentDate string `json:"PlannedPaymentDate,omitempty" xml:"PlannedPaymentDate,omitempty"`
 
 	// Total of invoice excluding taxes
-	SubTotal float32 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
+	SubTotal float64 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
 
 	// Total tax on invoice
-	TotalTax float32 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
+	TotalTax float64 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
 
 	// Total of Invoice tax inclusive (i.e. SubTotal + TotalTax). This will be ignored if it doesn’t equal the sum of the LineAmounts
-	Total float32 `json:"Total,omitempty" xml:"Total,omitempty"`
+	Total float64 `json:"Total,omitempty" xml:"Total,omitempty"`
 
 	// Total of discounts applied on the invoice line items
-	TotalDiscount float32 `json:"TotalDiscount,omitempty" xml:"-"`
+	TotalDiscount float64 `json:"TotalDiscount,omitempty" xml:"-"`
 
 	// Xero generated unique identifier for invoice
 	InvoiceID string `json:"InvoiceID,omitempty" xml:"InvoiceID,omitempty"`
@@ -88,16 +88,16 @@ type Invoice struct {
 	Overpayments *[]Overpayment `json:"Overpayments,omitempty" xml:"-"`
 
 	// Amount remaining to be paid on invoice
-	AmountDue float32 `json:"AmountDue,omitempty" xml:"-"`
+	AmountDue float64 `json:"AmountDue,omitempty" xml:"-"`
 
 	// Sum of payments received for invoice
-	AmountPaid float32 `json:"AmountPaid,omitempty" xml:"-"`
+	AmountPaid float64 `json:"AmountPaid,omitempty" xml:"-"`
 
 	// The date the invoice was fully paid. Only returned on fully paid invoices
 	FullyPaidOnDate string `json:"FullyPaidOnDate,omitempty" xml:"-"`
 
 	// Sum of all credit notes, over-payments and pre-payments applied to invoice
-	AmountCredited float32 `json:"AmountCredited,omitempty" xml:"-"`
+	AmountCredited float64 `json:"AmountCredited,omitempty" xml:"-"`
 
 	// Last modified date UTC format
 	UpdatedDateUTC string `json:"UpdatedDateUTC,omitempty" xml:"-"`
