+++ accounting/overpayment.go
@@ -32,13 +32,13 @@ type Overpayment struct {
 	LineItems []LineItem `json:"LineItems,omitempty" xml:"LineItems,omitempty"`
 
 	// The subtotal of the overpayment excluding taxes
-	SubTotal float32 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
+	SubTotal float64 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
 
 	// The total tax on the overpayment
-	TotalTax float32 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
+	TotalTax float64 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
 
 	// The total of the overpayment (subtotal + total tax)
-	Total float32 `json:"Total,omitempty" xml:"Total,omitempty"`
+	Total float64 `json:"Total,omitempty" xml:"Total,omitempty"`
 
 	// UTC timestamp of last update to the overpayment
 	UpdatedDateUTC string `json:"UpdatedDateUTC,omitempty" xml:"UpdatedDateUTC,omitempty"`
@@ -50,10 +50,10 @@ type Overpayment struct {
 	OverpaymentID string `json:"OverpaymentID,omitempty" xml:"OverpaymentID,omitempty"`
 
 	// The currency rate for a multicurrency overpayment. If no rate is specified, the XE.com day rate is used
-	CurrencyRate float32 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
+	CurrencyRate float64 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
 
 	// The remaining credit balance on the overpayment
-	RemainingCredit float32 `json:"RemainingCredit,omitempty" xml:"RemainingCredit,omitempty"`
+	RemainingCredit float64 `json:"RemainingCredit,omitempty" xml:"RemainingCredit,omitempty"`
 
 	// See Allocations
 	Allocations []Allocation `json:"Allocations,omitempty" xml:"Allocations,omitempty"`
